pyproject.toml
[project]
name = "effdetpai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.2.1",
    "antlr4-python3-runtime==4.9.3",
    "astunparse==1.6.3",
    "blinker==1.9.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "filelock==3.18.0",
    "flask==3.1.0",
    "flatbuffers==25.2.10",
    "fonttools==4.56.0",
    "fpdf>=1.7.2",
    "fsspec==2025.3.2",
    "gast==0.6.0",
    "google-pasta==0.2.0",
    "grpcio==1.71.0",
    "h5py==3.13.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "keras==3.9.1",
    "kiwisolver==1.4.8",
    "libclang==18.1.1",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.1",
    "mpmath==1.3.0",
    "namex==0.0.8",
    "networkx==3.4.2",
    "numpy==2.1.3",
    "omegaconf==2.3.0",
    "opencv-python==4.11.0.86",
    "opt-einsum==3.4.0",
    "optree==0.14.1",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "py-cpuinfo==9.0.0",
    "pycocotools==2.0.8",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==14.0.0",
    "safetensors==0.5.3",
    "scikit-learn>=1.6.1",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "setuptools==78.1.0",
    "six==1.17.0",
    "sympy==1.13.1",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.19.0",
    "termcolor==2.5.0",
    "timm==1.0.15",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.0",
    "tzdata==2025.2",
    "ultralytics==8.3.124",
    "ultralytics-thop==2.0.14",
    "urllib3==2.3.0",
    "webcolors==24.11.1",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wrapt==1.17.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
